fix(server): rename the watchdog per-layer test files and restore outer describe blocks
The prior split named the files domain.test.ts and application.test.ts
and dropped the outer describe wrapper in each file. Rename them to
policy.test.ts and use-cases.test.ts, next to the production file each
one tests, and restore the outer describe("domain"), describe("application"),
and describe("adapters") blocks so the layout and test names match the
merged file this change replaces.
